% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{abc_fit}
\alias{abc_fit}
\title{ABC fitting}
\usage{
abc_fit(obs_data, priors, iterations)
}
\arguments{
\item{obs_data}{observations on which to base the estimation from ZEST data}

\item{priors}{vector of ranges for prior distributions}

\item{iterations}{number of synthetic datasets to generate and compare to observed data}
}
\value{
list with abc objects that contain fitting statistics and generated datasets for each population
}
\description{
Conduct approximate bayesian computation using four proposed data-generated mechanisms to estimate worm burdens from observed egg burdens under different data-generating mechanisms
}
