% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{abc_post_pred}
\alias{abc_post_pred}
\title{ABC Posterior Predictive Check}
\usage{
abc_post_pred(abc_sims_list, yO, sim)
}
\arguments{
\item{abc_sims_list}{list containing abc objects to conduct posterior predicitve checks}

\item{yO}{observed data frame for matching shehia year and pop info}

\item{sim}{numeric indicating which simulation to pull from `abc_sims_list` and conduct posterior predictive checks on}
}
\value{
matrix with dim n_reps by number of summary statistics (3 for this analysis)
}
\description{
Generate and store summaries of posterior predictive check data
}
