% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_parameter_functions.R
\name{alpha_likelihood}
\alias{alpha_likelihood}
\title{Custom max likelihood function for dispersion parameter}
\usage{
alpha_likelihood(alpha, vec)
}
\arguments{
\item{alpha}{value of the inverse of the dispersion parameter, kappa}

\item{vec}{vector of individual egg or worm burden intensities}
}
\value{
estimate of the negative log likelihood
}
\description{
Estimates the likelihood of a vector of egg counts with negative binomial distribution 
fixed mean and inverse of the dispersion parameter, kappa, frequently denoted alpha
can be used in conjunction with an opimizer function to find max likelihood of the
dispersion parameter, implemented here in the `mle_kap` function
}
