% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{est_case3_meanW}
\alias{est_case3_meanW}
\title{Case 3 mean from parameters}
\usage{
est_case3_meanW(alphaS, betaS, meanC)
}
\arguments{
\item{alphaS}{gamma shape parameter for susceptibility}

\item{betaS}{gamma rate parameter for susceptibility}

\item{meanC}{mean cercarial exposure}
}
\value{
estimate of mean
}
\description{
Analytic estimate of mean as combination of neg. bin. cercarial exposures and gamma susceptibility
}
