% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{est_case3_varW}
\alias{est_case3_varW}
\title{Case 3 variance from parameters}
\usage{
est_case3_varW(alphaS, betaS, meanC, alphaC)
}
\arguments{
\item{alphaS}{gamma shape parameter for susceptibility}

\item{betaS}{gamma rate parameter for susceptibility}

\item{meanC}{mean cercarial exposure}

\item{alphaC}{inverse dispersion parameter of cercarial exposure}
}
\value{
estimate of variance
}
\description{
Analytic estimate of variance as combination of neg. bin. cercarial exposures and gamma susceptibility
}
