% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{est_dispW}
\alias{est_dispW}
\title{Moments estimator of dispersion parameter}
\usage{
est_dispW(meanW, varW)
}
\arguments{
\item{meanW}{mean worm burden}

\item{varW}{variance of worm burden}
}
\value{
estimate of dispersion parameter
}
\description{

}
