% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{gen_case12_pars}
\alias{gen_case12_pars}
\title{Generate priors for case 1/2 dataset generation}
\usage{
gen_case12_pars(
  iterations,
  mean_W_lo,
  mean_W_hi,
  disp_W_lo,
  disp_W_hi,
  h_lo,
  h_hi,
  r_lo,
  r_hi
)
}
\arguments{
\item{iterations}{numeric of number of simulations to run}

\item{h_lo}{lower range of daily mean egg release per mated pair}

\item{h_hi}{upper range of daily mean egg release per mated pair}

\item{r_lo}{lower range of daily egg release dispersion}

\item{r_hi}{upper range of daily egg release dispersion}

\item{mean_w_lo}{lower range of mean worm burden}

\item{mean_w_hi}{upper range of mean worm burden}

\item{disp_w_lo}{lower range of worm burden dispersion}

\item{disp_w_hi}{upper range of worm burden dispersion}
}
\value{
matrix of priors
}
\description{
Generate priors for ABC data generation for cases 1 and 2
}
