% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{gen_case3_data}
\alias{gen_case3_data}
\title{Generate synthetic case 3 dataset}
\usage{
gen_case3_data(pars, n)
}
\arguments{
\item{pars}{numeric vector with gamma shape and scale parameters for susceptibility distribution, mean cercarial exposure, and dispersion of cercarial exposure, mean eggs per mated pair, and dispersion of daily egg release}

\item{n}{number of individual human hosts}
}
\value{
vector of summary statistics to compare to observed in ABC setup
}
\description{
Generate worm pairs and resulting egg burdens for individuals assuming gamma distributed susceptibility and negative binomially distributed cercarial exposures
}
