% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{gen_case3_pars}
\alias{gen_case3_pars}
\title{Generate priors for case 3 dataset generation}
\usage{
gen_case3_pars(
  iterations,
  susc_shape_lo,
  susc_shape_hi,
  susc_rate_lo,
  susc_rate_hi,
  mean_C_lo,
  mean_C_hi,
  disp_C_lo,
  disp_C_hi,
  h_lo,
  h_hi,
  r_lo,
  r_hi,
  g_lo,
  g_hi
)
}
\arguments{
\item{iterations}{numeric of number of simulations to run}

\item{susc_shape_lo}{lower range of susceptibility shape parameter}

\item{susc_shape_hi}{upper range of susceptibility shape parameter}

\item{susc_rate_lo}{lower range of susceptibility rate parameter}

\item{susc_rate_hi}{upper range of susceptibility rate parameter}

\item{mean_C_lo}{lower range of mean cercarial exposure}

\item{mean_C_hi}{upper range of mean cercarial exposure}

\item{disp_C_lo}{lower range of cercarial exposure dispersion}

\item{disp_C_hi}{upper range of cercarial exposure dispersion}

\item{h_lo}{lower range of daily mean egg release per mated pair}

\item{h_hi}{upper range of daily mean egg release per mated pair}

\item{r_lo}{lower range of daily egg release dispersion}

\item{r_hi}{upper range of daily egg release dispersion}

\item{g_lo}{lower range of density dependent fecundity parameter}

\item{g_hi}{upper range of density dependent fecundity parameter}
}
\value{
matrix of priors
}
\description{
Generate priors for ABC data generation for case 3
}
