% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{get_abc_med_iqr}
\alias{get_abc_med_iqr}
\title{Median and IQR of abc posterior}
\usage{
get_abc_med_iqr(abc_pars, abc_weights)
}
\arguments{
\item{abc_pars}{vector of abc pars from fitting}

\item{abc_weights}{vector of weights returned from abc posterior correction}
}
\value{
vector with median and iqr of posterior
}
\description{
Utility function to return median and 25th and 75th quantiles of posterior distribution from ABC sim
}
