% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_bootstrapping_functions.R
\name{get_boot_df}
\alias{get_boot_df}
\title{Get data frame of observations for desired ZEST strata}
\usage{
get_boot_df(Island, Treatment, Population)
}
\arguments{
\item{Island}{Character indicating Island of observations (either "Pemba" or "Unguja")}

\item{Population}{Character indicating population of observations (either "Children or "Adults")}

\item{Intervention}{Character indicating intervention of observations (either "MDA" or "Snail" or "Behavior)}
}
\value{
Data frame of observations from full ZEST dataset meeting input stratifications
}
\description{

}
