% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_bootstrapping_functions.R
\name{get_boot_ests}
\alias{get_boot_ests}
\title{Get bootstrapped GEE estimates}
\usage{
get_boot_ests(df, boot_samps)
}
\arguments{
\item{df}{dataframe of observations to sample and fit GEE}

\item{boot_samps}{number of bootstrapped samples to take}
}
\value{
Vector of point estimates for bootstrapped dataframes
}
\description{

}
