% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_parameter_functions.R
\name{mean_var_agg_crctd}
\alias{mean_var_agg_crctd}
\title{Get aggregation parameter from observed data}
\usage{
mean_var_agg_crctd(vec)
}
\arguments{
\item{egg_vec}{vector of individual egg or worm burden estimates}
}
\value{
estimate of the aggregation parameter of the negative binomial dist'n
}
\description{
Estimates the aggregation parameter, `kappa`, from vector of observed values (egg or worm burden)
based on estimation of the mean and variance as: 
$$\kappa=\frac{\mu^2-\frac{\sigma^2}{n}}{\sigma^2-\mu}$$
}
