% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_parameter_functions.R
\name{mean_var_agg_par}
\alias{mean_var_agg_par}
\title{Estimate aggregation parameter from mean-variance relationship}
\usage{
mean_var_agg_par(vec)
}
\arguments{
\item{vec}{vector of individual egg or worm burden intensities}
}
\value{
estimate of the aggregation parameter
}
\description{
Estimate aggregation parameter from mean-variance relationship
}
