% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_parameter_functions.R
\name{mle_alpha}
\alias{mle_alpha}
\title{Finds max likelihood of the inverse of the dispersion parameter, alpha, dispersion parameter using `alpha_likelihood`}
\usage{
mle_alpha(vec)
}
\arguments{
\item{vec}{vector of individual egg or worm burden intensities}
}
\value{
max likelihood estimate of the dispersion parameter
}
\description{
Finds the maximum likelihood estimate of the inverse of the dispersion parameter,
kappa, often denoted alpha, given a vector of egg or worm burdens. Assumes max likelihood
estimate of the mean of the negative binomial distribution is the empirical mean of the vector
Returns the message from `optim` if convergence fails
}
