% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_parameter_functions.R
\name{mle_kap_se}
\alias{mle_kap_se}
\title{Finds max likelihood of the dispersion parameter using `alpha_likelihood`}
\usage{
mle_kap_se(vec)
}
\arguments{
\item{vec}{vector of individual egg or worm burden intensities}
}
\value{
standard error of the max likelihood estimate of the dispersion parameter
}
\description{
Finds the standard error of the maximum likelihood estimate of the inverse of the 
dispersion parameter,kappa, often denoted alpha, given a vector of egg or worm burdens. 
Uses sqrt of the inverse of the hessian returned by `optim` to estimate standard error of alpha
then transforms to estimate the standard error of kappa.
Assumes max likelihood estimate of the mean of the negative binomial distribution is the 
empirical mean of the vector
Returns the message from `optim` if convergence fails
}
