% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_parameter_functions.R
\name{mle_mean_par}
\alias{mle_mean_par}
\title{Estimate MLE mean parameter using `fitdistrplus::fitdist`}
\usage{
mle_mean_par(vec)
}
\arguments{
\item{vec}{vector of individual egg or worm burden intensities}
}
\value{
estimate of the mean parameter
}
\description{
Uses fitdist function from fitdistrplus package to estimate the mean
of vector of count data
}
