% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mating_probability_functions.R
\name{phi_wk_sep}
\alias{phi_wk_sep}
\title{Function to estimate mating probability for sexes distributed separately}
\usage{
phi_wk_sep(w, k)
}
\arguments{
\item{w}{mean worm burden in human population}

\item{k}{aggregation parameter of negative binomially distributed worms}
}
\value{
estimate of the Case II mating probability
}
\description{
derived in Anderson 1977 and Bradley and Anderson 1978 for schistosomes of opposite 
sexes drawn from separate negative binomial distributions (Case II: Distributed Separate)
This function requires `prob_j_pairs` function
}
