% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mating_probability_functions.R
\name{prob_j_pairs}
\alias{prob_j_pairs}
\title{Function to estimate the probability of an individual host having j mated pairs}
\usage{
prob_j_pairs(j, w, k)
}
\arguments{
\item{j}{number of mated pairs}

\item{w}{mean worm burden in human population}

\item{k}{aggregation parameter of negative binomially distributed worms}
}
\value{
estimate of the probability of j mated pairs in a human host
}
\description{
derived in Anderson 1977 and Bradley and Anderson 1978 for schistosomes of opposite 
sexes drawn from separate negative binomial distributions (Case II: Distributed Separate)
}
