% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_worm_estimation_functions.R
\name{quiet}
\alias{quiet}
\title{Quitely run}
\usage{
quiet(x)
}
\arguments{
\item{x}{command call}
}
\value{
silent result of command call
}
\description{
Run without outputting messages/warnings
}
